Fix(Screens): sorting the screens in separated folders
