--- src/screens/List.tsx
@@ -1,60 +1,0 @@
-import React from 'react';
-import { StyleSheet, FlatList } from 'react-native';
-import { StackNavigationProp } from '@react-navigation/stack';
-
-import colors from '../constants/colors';
-import { ListItem, ListSeparator } from '../components/List';
-import { loginStackParamList } from '../types/Types';
-
-const screens = [
-  {
-    title: 'Text',
-    subtitle: 'An example of using the Text.js components.',
-    target: 'TextDemo',
-  },
-  {
-    title: 'Form',
-    subtitle: 'An example of using the Form.js components.',
-    target: 'FormDemo',
-  },
-  {
-    title: 'Button',
-    subtitle: 'An example of using the Button.js components.',
-    target: 'ButtonDemo',
-  },
-];
-
-type Props = {
-  navigation: StackNavigationProp<loginStackParamList, 'List'>;
-};
-
-export const List = ({ navigation }: Props) => {
-  return (
-    <FlatList
-      style={styles.container}
-      data={screens}
-      keyExtractor={item => item.title}
-      renderItem={({ item }) => (
-        <ListItem
-          title={item.title}
-          subtitle={item.subtitle}
-          // @ts-ignore
-          // Disabling the next line because all the item.targets are valid - that data just
-          // isn't getting picked up by TypeScript
-          onPress={() => navigation.push(item.target)}
-        />
-      )}
-      ItemSeparatorComponent={ListSeparator}
-      ListHeaderComponent={ListSeparator}
-      ListFooterComponent={ListSeparator}
-    />
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: colors.background,
-    paddingVertical: 20,
-  },
-});
