+++ src/screens/account/Account.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { View } from 'react-native';
+import { Text } from '../../components/Text';
+import { styles } from './AccountStyles';
+
+export const Account = () => (
+  <View style={styles.container}>
+    <Text type="header">ACCOUNT</Text>
+  </View>
+);
