+++ src/screens/addFunds/AddFunds.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+import { styles } from './AddFundsStyles';
+
+export const AddFunds = () => (
+  <View style={styles.container}>
+    <Text>ADDFUNDS</Text>
+  </View>
+);
