+++ src/screens/home/Home.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { View } from 'react-native';
+import { Text } from '../../components/Text';
+import { styles } from './HomeStart';
+
+export const Home = () => (
+  <View style={styles.container}>
+    <Text type="header">HOME</Text>
+  </View>
+);
