+++ src/screens/index.ts
@@ -1,0 +1,17 @@
+import { Start } from '../screens/start/Start';
+import { Login } from '../screens/login/Login';
+import { Register } from '../screens/register/Register';
+import { Home } from '../screens/home/Home';
+import { Transfer } from '../screens/transfer/Transfer';
+import { Account } from '../screens/account/Account';
+import { AddFunds } from './addFunds/AddFunds';
+
+export const config = {
+  Start,
+  Login,
+  Register,
+  Home,
+  AddFunds,
+  Transfer,
+  Account,
+};
