+++ src/screens/transfer/Transfer.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { View } from 'react-native';
+import { Text } from '../../components/Text';
+import { styles } from './TransferStyles';
+
+export const Transfer = () => (
+  <View style={styles.container}>
+    <Text type="header">TRANSFER</Text>
+  </View>
+);
